requirements.txt
Package           Version
----------------- -------
annotated-types   0.6.0
anyio             4.2.0
cffi              1.16.0
click             8.1.7
colorama          0.4.6
cryptography      42.0.1
fastapi           0.109.0
h11               0.14.0
httptools         0.6.1
idna              3.6
jose              1.0.0
pip               22.3
pycparser         2.21
pydantic          2.5.3
pydantic_core     2.14.6
PyJWT             2.8.0
python-dotenv     1.0.1
PyYAML            6.0.1
setuptools        65.5.0
sniffio           1.3.0
starlette         0.35.1
typing_extensions 4.9.0
uvicorn           0.27.0
watchfiles        0.21.0
websockets        12.0
